backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
python-multipart
sqlalchemy
psycopg2-binary
fastapi-mail
python-dotenv
alembic
pydantic 
pydantic[email]
bcrypt